* src/com/fluendo/jheora/DCTDecode.java: (DCTDecode.....CopyRecon), (DCTDecode.....CopyNotRecon), (DCTDecode.....ReconRefFrames): Move UpdateUMVBorder after the loop functions. * src/com/fluendo/jheora/iDCT.java: (iDCT.IDctSlow), (iDCT.IDct10): Update IDCT 16 bit truncating updates. * src/com/fluendo/player/Cortado.java: (Cortado), (Cortado.getParameterInfo), (Cortado.init), (Cortado.getGraphics), (Cortado.paint): Remove framerate property Added statusHeight property to configure the size of the status area. * src/com/fluendo/player/Status.java: (Status), (Status.paintStop), (Status.paintMessage), (Status.paintBuffering), (Status.paint), (Status.setHavePercent), (Status.setSeekable), (Status.setState), (Status.intersectButton1), (Status.intersectButton2), (Status.intersectSeek), (Status.mouseReleased), (Status.mouseDragged): Make statusHeight configurable. Make the slider handle a bit bigger.
git-svn-id: https://code.fluendo.com/flumotion/svn/cortado/trunk@3996 2b0047a9-a6d8-0310-accf-f7200b2a168c

+++ src/com/fluendo/jheora/iDCT.java
@@ -87,14 +87,14 @@ public final class iDCT
         t2 = (xC5S3 * ip[3 + off]) >> 16;
         _D = t1 - t2;
 
-        _Ad = (xC4S4 * (_A - _C)) >> 16;
-        _Bd = (xC4S4 * (_B - _D)) >> 16;
+        _Ad = (xC4S4 * (short)(_A - _C)) >> 16;
+        _Bd = (xC4S4 * (short)(_B - _D)) >> 16;
 
         _Cd = _A + _C;
         _Dd = _B + _D;
 
-        _E = (xC4S4 * (ip[0 + off] + ip[4 + off])) >> 16;
-        _F = (xC4S4 * (ip[0 + off] - ip[4 + off])) >> 16;
+        _E = (xC4S4 * (short)(ip[0 + off] + ip[4 + off])) >> 16;
+        _F = (xC4S4 * (short)(ip[0 + off] - ip[4 + off])) >> 16;
 
         t1 = (xC2S6 * ip[2 + off]) >> 16;
         t2 = (xC6S2 * ip[6 + off]) >> 16;
@@ -150,14 +150,14 @@ public final class iDCT
         t2 = (xC5S3 * ip[3*8 + off]) >> 16;
         _D = t1 - t2;
 
-        _Ad = (xC4S4 * (_A - _C)) >> 16;
-        _Bd = (xC4S4 * (_B - _D)) >> 16;
+        _Ad = (xC4S4 * (short)(_A - _C)) >> 16;
+        _Bd = (xC4S4 * (short)(_B - _D)) >> 16;
 
         _Cd = _A + _C;
         _Dd = _B + _D;
 
-        _E = (xC4S4 * (ip[0*8 + off] + ip[4*8 + off])) >> 16;
-        _F = (xC4S4 * (ip[0*8 + off] - ip[4*8 + off])) >> 16;
+        _E = (xC4S4 * (short)(ip[0*8 + off] + ip[4*8 + off])) >> 16;
+        _F = (xC4S4 * (short)(ip[0*8 + off] - ip[4*8 + off])) >> 16;
 
         t1 = (xC2S6 * ip[2*8 + off]) >> 16;
         t2 = (xC6S2 * ip[6*8 + off]) >> 16;
@@ -246,8 +246,8 @@ public final class iDCT
         _C = (xC3S5 * ip[3 + off]) >> 16;
         _D = (xC5S3 * ip[3 + off]) >> 16;
 
-        _Ad = (xC4S4 * (_A - _C)) >> 16;
-        _Bd = (xC4S4 * (_B - _D)) >> 16;
+        _Ad = (xC4S4 * (short)(_A - _C)) >> 16;
+        _Bd = (xC4S4 * (short)(_B - _D)) >> 16;
 
         _Cd = _A + _C;
         _Dd = _B + _D;
@@ -291,8 +291,8 @@ public final class iDCT
         _C = (xC3S5 * ip[3*8 + off]) >> 16;
         _D = (xC5S3 * ip[3*8 + off]) >> 16;
 
-        _Ad = (xC4S4 * (_A - _C)) >> 16;
-        _Bd = (xC4S4 * (_B - _D)) >> 16;
+        _Ad = (xC4S4 * (short)(_A - _C)) >> 16;
+        _Bd = (xC4S4 * (short)(_B - _D)) >> 16;
 
         _Cd = _A + _C;
         _Dd = _B + _D;
